Move kinvey-nativescript-sdk package
