--- old_packages/kinvey-nativescript-sdk/package.json
@@ -1,69 +1,0 @@
-{
-  "private": true,
-  "version": "3.9.3-beta.7",
-  "name": "kinvey-nativescript-sdk",
-  "description": "Kinvey NativeScript SDK for developing NativeScript applications.",
-  "author": "Kinvey, Inc.",
-  "homepage": "http://www.kinvey.com",
-  "license": "Apache-2.0",
-  "main": "kinvey-nativescript-sdk.min.js",
-  "typings": "kinvey.d.ts",
-  "devDependencies": {
-    "copy-webpack-plugin": "^4.2.3",
-    "fs-extra": "4.0.2",
-    "json-loader": "^0.5.4",
-    "ts-loader": "^3.1.1",
-    "kinvey": "^3.9.3-beta.7",
-    "kinvey-acl": "^3.9.3-beta.7",
-    "kinvey-aggregation": "^3.9.3-beta.7",
-    "kinvey-datastore": "^3.9.3-beta.7",
-    "kinvey-endpoint": "^3.9.3-beta.7",
-    "kinvey-errors": "^3.9.3-beta.7",
-    "kinvey-filestore": "^3.9.3-beta.7",
-    "kinvey-html5": "^3.9.3-beta.7",
-    "kinvey-html5-cache": "^3.9.3-beta.7",
-    "kinvey-html5-http": "^3.9.3-beta.7",
-    "kinvey-html5-popup": "^3.9.3-beta.7",
-    "kinvey-identity": "^3.9.3-beta.7",
-    "kinvey-live": "^3.9.3-beta.7",
-    "kinvey-log": "^3.9.3-beta.7",
-    "kinvey-metadata": "^3.9.3-beta.7",
-    "kinvey-query": "^3.9.3-beta.7",
-    "kinvey-request": "^3.9.3-beta.7",
-    "kinvey-user": "^3.9.3-beta.7",
-    "rimraf": "2.6.2",
-    "uglifyjs-webpack-plugin": "1.0.1",
-    "webpack": "^2.5.1",
-    "es6-promise": "4.1.0",
-    "events": "1.1.1",
-    "tns-core-modules": "3.0.1",
-    "tns-platform-declarations": "3.0.1",
-    "tslint": "5.7.0",
-    "tslint-config-airbnb": "5.3.0",
-    "typescript": "2.3.4"
-  },
-  "scripts": {
-    "prebuild": "rimraf dist",
-    "build": "npm run build:android && npm run build:ios",
-    "build:android": "webpack --config webpack.config.js --env.android",
-    "build:ios": "webpack --config webpack.config.js --env.ios",
-    "lint": "tslint **/*.ts --exclude **/node_modules/**/* --exclude **/*.d.ts",
-    "release": "npm run build && ./scripts/push.sh",
-    "postrelease": "rimraf tmp"
-  },
-  "files": [
-    "kinvey-nativescript-sdk.ios.min.js",
-    "kinvey-nativescript-sdk.ios.min.js.map",
-    "kinvey-nativescript-sdk.android.min.js",
-    "kinvey-nativescript-sdk.android.min.js.map",
-    "kinvey.d.ts",
-    "platforms/android/include.gradle",
-    "platforms/ios/Podfile"
-  ],
-  "nativescript": {
-    "platforms": {
-      "ios": "3.0.0",
-      "android": "3.0.0"
-    }
-  }
-}
