+++ package.json
@@ -39,6 +39,12 @@
     "nock": "~9.1.3",
     "proxyquire": "~1.8.0",
     "rimraf": "~2.6.2",
+    "tns-core-modules": "3.0.1",
+    "tns-platform-declarations": "3.0.1",
+    "ts-loader": "~3.2.0",
+    "tslint": "5.7.0",
+    "tslint-config-airbnb": "5.3.0",
+    "typescript": "2.3.4",
     "uglifyjs-webpack-plugin": "~1.0.1",
     "webpack": "~3.9.1"
   }
