+++ packages/kinvey-nativescript-sdk/.travis.yml
