+++ packages/kinvey-nativescript-sdk/LICENSE
