+++ packages/kinvey-nativescript-sdk/README.md
