+++ packages/kinvey-nativescript-sdk/mangle-excludes.js
