+++ packages/kinvey-nativescript-sdk/platforms/android/include.gradle
