+++ packages/kinvey-nativescript-sdk/platforms/ios/Podfile
