+++ packages/kinvey-nativescript-sdk/scripts/push.sh
