+++ src/nativescript/cache/sqlite.ts
@@ -1,5 +1,5 @@
-import { isDefined } from 'kinvey-utils/object';
-import { NotFoundError } from 'kinvey-errors';
+import { isDefined } from '../../core/utils';
+import { NotFoundError } from '../../core/errors';
 const nativeScriptSQLite = require('nativescript-sqlite');
 
 const masterCollectionName = 'sqlite_master';
