+++ src/nativescript/filestore/filestore.android.ts
