+++ src/nativescript/filestore/filestore.d.ts
