+++ src/nativescript/filestore/package.json
