+++ src/nativescript/popup/package.json
