+++ src/nativescript/popup/popup.android.ts
