+++ src/nativescript/popup/popup.d.ts
