+++ src/nativescript/popup/popup.ios.ts
