+++ src/nativescript/push/common.js
