+++ src/nativescript/push/package.json
