+++ src/nativescript/push/push.ios.js
