+++ src/nativescript/secure/package.json
