+++ src/nativescript/secure/secure.android.ts
