+++ src/nativescript/secure/secure.d.ts
