+++ src/nativescript/secure/secure.ios.ts
